refactor(web): extract ConnectionForm defaults into schema files
Extracts form defaults from ConnectionForm.tsx into separate
*FieldsSchema.ts files organized by field group:
  - bondFieldsSchema.ts (bond-specific fields)
  - bridgeFieldsSchema.ts (bridge-specific fields)
  - commonFieldsSchema.ts (common fields like name, type, binding)
  - ipFieldsSchema.ts (IPv4/IPv6 settings)
  - connectionSchema.ts (composition + formOptions)

Also moves FormIpMode, BridgeStpMode, and ADDRESS_REQUIRED_MODES
constants to their respective schema files to avoid circular imports.

This modular structure:
  - Matches the validation file organization
  - Makes defaults easier to locate and maintain
  - Prepares for Valibot schema addition (next commit)
  - Provides a clear pattern for adding new connection types
