+++ web/src/components/network/connection-form/BindingModeSelector.tsx
@@ -21,7 +21,7 @@
  */
 
 import React from "react";
-import { connectionFormOptions } from "./ConnectionForm";
+import { connectionFormOptions } from "./connectionSchema";
 import { withForm } from "~/hooks/form";
 import { _, N_ } from "~/i18n";
