+++ web/src/components/network/connection-form/BridgeFields.tsx
@@ -25,7 +25,8 @@ import { sprintf } from "sprintf-js";
 import Interpolate from "~/components/core/Interpolate";
 import NestedContent from "~/components/core/NestedContent";
 import LabelText from "~/components/form/LabelText";
-import { connectionFormOptions, BridgeStpMode } from "./ConnectionForm";
+import { connectionFormOptions } from "./connectionSchema";
+import { BridgeStpMode } from "./bridgeFieldsSchema";
 import { withForm } from "~/hooks/form";
 import { useDevices } from "~/hooks/model/system/network";
 import { _, N_, formatList } from "~/i18n";
